requirements.txt
criteria-etl==1.0a5
jupyter==1.*
matplotlib==3.*
pandas==1.*
scikit-learn==0.*
xlrd<=1.2
pyreadstat==1.0.*
unidecode==1.1.*
statsmodels==0.12.*
openpyxl==3.0.*
XlsxWriter==1.3.*
boto3==1.16.*
seaborn==0.11.*
python-dotenv==0.12.*
geopy==2.1.*
tqdm==4.61.*
textdistance==4.2.*
